Create pagination for register and login
